requirements.txt
google-genai>=1.0.0
cryptography>=41.0
python-dotenv>=1.0.0
PyGithub>=2.0.0
